First try for the Armbian docker configuration module
Currently, the only thing tested is how to popup a selection UI
if no arguments are passed, in order to fill the arguments.

The design is still pretty much done backwards at the moment,
focusing too much on the program parameters, instead of adapting
the parameters to a configuration file.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>