Not inspired
I forgot the fact that I still have no clear way
to detect if the UI can be run or not.

But the biggest issue is that I'm clearly not
inspired on this project.
I still have no clear idea where this is going,
even though there's a roadmap.

QT is sucking out the will I'd like to put on this
project, due to a FUCKING HORRENDOUS toolset.
The more I'm using it, the more I feel that QML is
just "that garbage that QT doesn't know what to do
with". This reminds that I'll have to make a converter
from Godot Engine, or whatever over UI maker that
makes a bit of sense in this era.

So yeah...

The best way to tackle this would be to have a UI
focused on presenting all the available docker images,
while also providing extended view and toolset for
installed images.
This can be done. However, then I'm still wondering
how to generalize this afterwards.
Hence the 'I feel I'm alone on this project that is
not mine'.

So, unless there's some real movement on the project,
this might be one of the last commit.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>